fix profile links
