+++ web/index.html
@@ -3,12 +3,8 @@
 
 <head>
   <title>Revive Launcher</title>
-  <script src="https://unpkg.com/vue"></script>
-  <script src="https://unpkg.com/vue-resource"></script>
   <link rel="stylesheet" type="text/css" href="./css/reset.css" />
   <link rel="stylesheet" type="text/css" href="./css/main.css" />
-  <script src="./js/main.js"></script>
-  <script src="./js/UserController.js"></script>
 </head>
 
 <body>
@@ -85,6 +81,49 @@
       </ul>
     </section>
   </main>
+  <template id="user-template">
+    <div class="user" :onClick='"onProfileClick("+user.id+")"' :data-status=" user.online? 'online': 'offline' ">
+      <figure class="avatar "><img alt="avatar " :src=" 'https://www.gravatar.com/avatar/'+user.avatar " draggable="false " /> </figure>
+      <span class="username ">{{ user.username }}</span>
+    </div>
+  </template>
+  <template id="soldier-select-modal-template ">
+    <transition name="modal ">
+      <div class="modal-mask ">
+        <div class="modal-wrapper ">
+          <div class="modal-container ">
+
+            <div class="modal-header ">
+              <slot name="header ">
+                Select Soldier
+              </slot>
+            </div>
+
+            <div class="modal-body ">
+              <slot name="body ">
+                <select v-model:currentSoldier>
+              <option v-for="soldier in soldiers " :value="soldier.nickname ">{{soldier.nickname}}</option>
+            </select>
+              </slot>
+            </div>
+
+            <div class="modal-footer ">
+              <slot name="footer ">
+                <button class="modal-default-button " @click="$emit( 'close') ">
+                Launch
+              </button>
+              </slot>
+            </div>
+          </div>
+        </div>
+      </div>
+    </transition>
+  </template>
+  <script src="https://unpkg.com/vue "></script>
+  <script src="https://unpkg.com/vue-resource "></script>
+  <script src="./js/main.js "></script>
+  <script src="./js/components.js "></script>
+  <script src="./js/user.js "></script>
 </body>
 
 </html>
