+++ web/js/components.js
@@ -1,0 +1,7 @@
+Vue.component('user', {
+    props: ['user'],
+    template: "#user-template"
+});
+Vue.component('soldier-select-modal', {
+    template: "#soldier-select-modal-template"
+})
