+++ web/js/soldier.js
@@ -1,0 +1,9 @@
+
+new Vue({
+    el: '#games',
+    data: {
+        showModal: false,
+        currentSoldier: false,
+        soldiers: getCatchedUser().soldiers
+    }
+})
